agents/sentiment_agent/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
httpx>=0.24.1
pydantic>=2.3.0
python-dotenv>=1.0.0
crewai>=0.27.0
requests>=2.28.0
google-generativeai
mcp>=1.0.0
nest_asyncio>=1.5.6
redditwarp>=1.3.0
# uuid is a standard library module and doesn't need to be installed 